+++ vitest.config.ts
@@ -4,6 +4,10 @@ import { defineConfig } from "vitest/config";
 const migrations = await readD1Migrations("migrations");
 
 export default defineConfig({
+  test: {
+    include: ["test/**/*.test.ts"],
+    exclude: ["test/._*.test.ts"],
+  },
   plugins: [
     cloudflareTest({
       wrangler: { configPath: "./wrangler.jsonc" },